# perf(rome_rowan): `SyntaxNodeText` improvements