This PR improves the performance of `SyntaxNodeText` by using the `token_at_offset` method rather than traversing over all tokens in the tree until it find the first token that has an overlapping range. Using `token_at_offset` has better performance because it only traverses into nodes that have an overlapping range.

Example: Getting the text for `+ e` in ``a + b + c + d + e`

* Old implementation: Traverses through the whole tree until it reaches the `+` token
* New: Finds the `+` on the first level, then traverses into `e`

Doing this improvement I discovered that `noShoutyConstants` uses `SyntaxNodeText` over `SyntaxTokenText`. `SyntaxTokenText` has the better performance because it isn't necessary to find the first token with an overlapping range. So, I refactored the implementation to use `SyntaxTokenText` instead

 ## Performance

 ```
 group                                    main                                    text
 -----                                    ----                                    ----
 analyzer/css.js                          1.06  1238.1±109.59µs     9.4 MB/sec    1.00   1167.7±6.84µs     9.9 MB/sec
 analyzer/index.js                        1.03      3.6±0.36ms     9.2 MB/sec     1.00      3.5±0.30ms     9.4 MB/sec
 analyzer/lint.ts                         1.04      2.1±0.07ms    19.6 MB/sec     1.00      2.0±0.01ms    20.4 MB/sec
 analyzer/parser.ts                       1.05      4.3±0.35ms    11.4 MB/sec     1.00      4.1±0.17ms    12.0 MB/sec
 analyzer/router.ts                       1.00      2.8±0.14ms    21.9 MB/sec     1.02      2.9±0.02ms    21.5 MB/sec
 analyzer/statement.ts                    1.00      3.8±0.17ms     9.3 MB/sec     1.01      3.8±0.12ms     9.3 MB/sec
 analyzer/typescript.ts                   1.22      7.4±1.25ms     7.3 MB/sec     1.00      6.1±0.46ms     9.0 MB/sec
```

## Test Plan

I added new doc tests and integration tests for the changes in `rome_rowan` and `rome_text_size`